pyproject.toml
[build-system]
requires = ["setuptools>=77.0"]
build-backend = "setuptools.build_meta"

[project]
name = "my-py-template"
version = "0.1.0"
description = "Python Starter Template: A minimal Python project template with a pre-configured pyproject.toml to get started quickly. Includes configuration stubs for: mypy, pytest, pre-commit, Github CI/CD. This repository is intended as a clean slate for new projects: lightweight, flexible, and easy to extend."
authors = [
    { name = "Giovanni Minuto", email = "minutogiovanni@gmail.com" }
]
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "pyspark==4.0.1",
    "delta-spark==4.0.0",
    "types-requests",
    "nbconvert",        # convert notebooks
    "ipykernel",        # Jupyter kernel support
]

[project.optional-dependencies]
dev = [
    "mypy",
    "black",
    "ruff",
    "pre-commit",
]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
disallow_untyped_defs = true
check_untyped_defs = true
strict_optional = true
warn_unused_ignores = true
warn_return_any = true
warn_redundant_casts = true
files = ["pinn_database/src"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
select = ["E", "F", "I", "Q"]
extend-ignore = ["F841", "E501"]
line-length = 100
indent-width = 4

[tool.ruff.isort]
required-imports = ["from __future__ import annotations"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.mccabe]
max-complexity = 15

[tool.ruff.flake8-quotes]
docstring-quotes = "double"
